Move apply code into own module
