+++ api/src/Hanabi.hs
@@ -1,7 +1,9 @@
 module Hanabi
   ( mkState
   , app
+  , applyWithInvalid
   ) where
 
 import Hanabi.Types
 import Hanabi.Api
+import Hanabi.Apply
