+++ api/src/Hanabi/Api.hs
@@ -9,6 +9,7 @@ module Hanabi.Api
 
 import Hanabi.Types
 import Hanabi.Json
+import Hanabi.Apply
 
 import           Network.Wai.Middleware.Cors          (cors, corsRequestHeaders,
                                                        simpleCorsResourcePolicy)
